+++ include/operator.hpp
@@ -1,0 +1,22 @@
+#ifndef CAST_OPERATOR_
+#define CAST_OPERATOR_
+
+#include "network_component.hpp"
+
+namespace cast {
+
+
+
+/**
+* Single-input, single-output network component.
+*
+* Includes `Layer` and `ActivationFunction` as its subclasses.
+*/
+class Operator : public NetworkComponent {
+};
+
+
+
+
+}
+#endif