+++ include/optimizer.hpp
@@ -4,6 +4,7 @@
 
 #include "layer.hpp"
 
+#include <initializer_list>
 #include <xtensor/generators/xbuilder.hpp>
 #include <xtensor/io/xio.hpp>
 
@@ -21,7 +22,31 @@ namespace cast {
 * Updates weights of a network's layers
 */
 class Optimizer {
+protected:
+    /**
+    * Stores hyperparameters of the optimizer
+    */
+    std::vector<double> hyperparams_;
+
 public:
+
+    /**
+    * @return deep copy of the optimizer's pointer. The new pointer cannot be used to modify the original.
+    */
+    virtual std::shared_ptr<Optimizer> shared_ptr_deep_copy() const = 0;
+
+    /**
+    * @return optimizer's hyperparameters
+    */
+    virtual std::vector<double> hyperparameters() const {
+        return hyperparams_;
+    }
+
+    /**
+    * Sets the hyperparameters to `new_hyperparams`.
+    * @param new_hyperparams hyperparameters to set
+    */
+    virtual void set_hyperparameters(std::initializer_list<double> new_hyperparams) = 0;
 
     /**
     * Loads the optimizer with all information needed for training.
@@ -50,16 +75,6 @@ class SGD : public Optimizer {
 private:
 
     /**
-    * Dictates speed of convergence
-    */
-    double learning_rate_;
-
-    /**
-    * Dictates momentum
-    */
-    double momentum_coefficient_;
-
-    /**
     * Operators that this optimizer improves
     */
     std::vector<std::shared_ptr<NetworkComponent>> operators_;
@@ -74,17 +89,70 @@ private:
 
 public:
     /**
+    * SGD hyperparameter indices
+    */
+    enum HyperparamIndices {
+        /**
+        * Speed of convergence
+        */
+        LearningRate = 0,
+
+        /**
+        * Momentum
+        */
+        MomentumCoefficient = 1
+    };
+
+
+    /**
     * Creates a new SGD optimizer with initial learning rate `initial_lr`
     * @param initial_lr initial learning rate to use. Positive
     * @param initial_momentum_coeff initial momentum coefficient to use. Non-negative
     */
-    SGD(double initial_lr, double initial_momentum_coeff) : learning_rate_(initial_lr), momentum_coefficient_(initial_momentum_coeff) {
+    SGD(double initial_lr, double initial_momentum_coeff) {
         str_assert(initial_lr > 0, "Initial learning rate (" + std::to_string(initial_lr) + ") must be positive");
         str_assert(initial_momentum_coeff >= 0, "Initial momentum coefficient (" + std::to_string(initial_momentum_coeff) + ") must be non-negative");
+
+        hyperparams_.push_back(initial_lr);
+        hyperparams_.push_back(initial_momentum_coeff);
     }
 
+    /**
+    * @return deep pointer copy of this SGD object
+    */
+    std::shared_ptr<Optimizer> shared_ptr_deep_copy() const override {
+        return std::make_shared<SGD>(*this);
+    }
 
+    
+    /**
+    * @return learning rate used by this optimizer
+    */
+    double learning_rate() const {
+        return hyperparams_[LearningRate];
+    }
 
+    /**
+    * @return momentum coefficient used by this optimizer
+    */
+    double momentum_coefficient() const {
+        return hyperparams_[MomentumCoefficient];
+    }
+
+    /**
+    * Sets the optimizer's learning rate to `new_hyperparams[0]`, and the momentum coefficient to `new_hyperparams[1]`.
+    * @param new_hyperparams new hyperparameters. Of length 2. Learning rate is positive, momentum coeff is non-negative
+    */
+    void set_hyperparameters(std::initializer_list<double> new_hyperparams) override {
+        std::vector<double> new_hyperparams_vec = new_hyperparams;
+        str_assert(new_hyperparams_vec.size() == 2, "New hyperparameter list must be of length 2");
+        str_assert(new_hyperparams_vec[LearningRate] > 0, "Learning rate must be positive; received " + std::to_string(new_hyperparams_vec[LearningRate]));
+        str_assert(new_hyperparams_vec[MomentumCoefficient] >= 0, "Momentum coefficient must be non-negative; received " + std::to_string(new_hyperparams_vec[MomentumCoefficient]));
+
+        hyperparams_ = new_hyperparams;
+    }
+
+    
     /**
     * Loads the SGD optimizer with all information needed for training, taken from `operators`.
     * @param operators operators to optimize. Non-empty, and no element can be `nullptr`
@@ -140,7 +208,7 @@ public:
             for(int32_t i = 0; i < current_layer->parameters().size(); i++) {
 
                 //v = momentum * v + learning_rate * gradient
-                vels[i] = momentum_coefficient_ * vels[i] + learning_rate_ * grads[i];
+                vels[i] = hyperparams_[MomentumCoefficient] * vels[i] + hyperparams_[LearningRate] * grads[i];
 
                 //param = param - v
                 params[i] -= vels[i];